# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GovCore Platform API — Identity Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /attachments:
    get:
      operationId: listAttachments
      tags: [attachments]
      summary: List uploaded attachments
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: attachedToType
          in: query
          schema:
            type: string
        - name: attachedToId
          in: query
          schema:
            type: string
        - name: documentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of attachments (without file content)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
    post:
      operationId: createAttachment
      tags: [attachments]
      summary: Upload a file and get back the UUID a form field stores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentInput"
      responses:
        "201":
          description: Attachment stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /attachments/{uuid}:
    get:
      operationId: getAttachment
      tags: [attachments]
      summary: Get an attachment's metadata and content
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The attachment with its base64 content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentWithContent"
        "404":
          description: Attachment not found
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, tenantId, name, code, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [tenantId, name, code]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: string
        parentId:
          type: integer
        status:
          type: string
          enum: [active, inactive]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        parentId:
          type: integer
        status:
          type: string
          enum: [active, inactive]

    # ── USER ────────────────────────────────────────────────────────────────────
...
    Attachment:
      type: object
      required: [id, uuid, tenantId, fileName, mimeType, sizeBytes, storageProvider, createdAt]
      properties:
        id:
          type: integer
        uuid:
          type: string
          description: The value a form file field stores in its submission
        tenantId:
          type: integer
        documentId:
          type: integer
          nullable: true
        attachedToType:
          type: string
          nullable: true
        attachedToId:
          type: string
          nullable: true
        fieldKey:
          type: string
          nullable: true
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        storageProvider:
          type: string
        storageKey:
          type: string
          nullable: true
        contentHash:
          type: string
          nullable: true
        uploadedByUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    AttachmentWithContent:
      allOf:
        - $ref: "#/components/schemas/Attachment"
        - type: object
          properties:
            content:
              type: string
              nullable: true
              description: Base64 file content
    AttachmentInput:
      type: object
      required: [tenantId, fileName, mimeType, content]
      properties:
        tenantId:
          type: integer
        fileName:
          type: string
        mimeType:
          type: string
        content:
          type: string
          description: Base64 file content
        attachedToType:
          type: string
          description: e.g. "form_submission"
        attachedToId:
          type: string
        fieldKey:
          type: string
          description: The form field this file satisfies
        documentId:
          type: integer
...